requirements.txt
keras~=3.3.3
scikit-learn~=1.5.0
tensorflow~=2.16.1
mysql-connector-python~=8.4.0
python-dotenv~=1.0.1
pandas~=2.2.2